+++ src/components/OldApp.tsx
@@ -1,7 +1,7 @@
 import { Footer } from '@/components/Footer';
 import { Header } from '@/components/Header';
 import { Projects } from '@/components/Projects';
-import { Theme } from '@/components/Theme';
+import { Theme } from '@/components/OldTheme';
 import { Title } from '@/components/Title';
 
 import '@/components/app.css';
